requirements-dev.txt
-r requirements.txt
django-debug-toolbar
django-elasticsearch-debug-toolbar
ipdb
coverage==4.4.1
coveralls==1.2.0
